File scope namespace
